+++ examples/index.php
@@ -12,7 +12,7 @@ $router = new Router(__DIR__ . "/views", "demo");
  * @route /
  */
 $router->get('/', function ($req, $res) {
-    return @$res->send("<h1>Hello World</h1> </br> Source: static GET </br> Query(name): {$req->query('name')}")->status(200);
+    return $res->send("<h1>Hello World</h1> <br/> Source: static GET </br> Query(name): {$req->query('name')}")->status(200);
 });
 
 
@@ -20,8 +20,8 @@ $router->get('/', function ($req, $res) {
  * @desc Serving a view/using a controller
  * @route /use
  */
-$router->get('/use', function ($req, $res) {
-    return $res->use("second.php")->status(200);
+$router->get('/render', function ($req, $res) {
+    return $res->render("second.php")->status(200);
 });
 
 /**
@@ -41,7 +41,7 @@ $router->get('/dynamic/:id', function ($req, $res) {
 });
 
 /**
- * @desc [GET] Nested dynamic route
+ * @desc [GET] Mixed dynamic route
  * @route /dynamic/:id/nested
  */
 $router->get('/dynamic/:id/nested', function ($req, $res) {
@@ -49,12 +49,43 @@ $router->get('/dynamic/:id/nested', function ($req, $res) {
 });
 
 /**
- * @desc [GET] Deeper dynamic route
+ * @desc [GET] Nested dynamic route
  * @route /dynamic/:id/:name
  */
 $router->get('/dynamic/:id/:name', function ($req, $res) {
     return $res->send("<h1>Hello World</h1> </br> Source: dynamic nested GET! </br> Param(ID): {$req->params("id")} </br> Param(Name): {$req->params("name")}")->status(200);
 });
+
+/**
+ * @desc Using middlewares
+ * @route /middleware
+ */
+$router->get('/middleware', function ($req, $res) {
+    $req->append("name", "John");// Appending data to the request object
+}, function ($req, $res) {
+    $req->append("age", 16); // Appending more data to the request object
+    $res->send("From middleware 2 <br/> 
+            Name from previous middleware: {$req->data["name"]} 
+            </br> ------ </br>");
+}, function ($req, $res) {
+    return $res->send("FROM FINAL CALLBACK </br> 
+                Name: {$req->data["name"]}</br> 
+                Age: {$req->data["age"]}")->status(200);
+});
+
+
+/**
+ * @desc Using chained routes
+ * @route /chained
+ */
+$router->route("/chained")
+    ->get(function ($req, $res) {
+    return $res->send("GET - Chained!")->status(200);
+    })
+    ->post(function ($req, $res) {
+    return $res->send("GET - Chained!")->status(200);
+    });
+
 
 /**
  * @desc [GET] Redirecting to another route
@@ -69,7 +100,7 @@ $router->get("/redirect", function ($req, $res) {
  * @route /
  */
 $router->post('/', function ($req, $res) {
-    return $res->send(["name" => $req->body("name"), "method" => "POST"])->status(200);
+    return $res->send(["name" => $req->body("name"), "method" => "POST"]);
 });
 
 /**
@@ -77,7 +108,7 @@ $router->post('/', function ($req, $res) {
  * @route /:id
  */
 $router->post('/:id', function ($req, $res) {
-    return $res->send(["id" => $req->params("id")])->status(200);
+    return $res->send(["id" => $req->params("id")]);
 });
 
 /**
@@ -85,7 +116,7 @@ $router->post('/:id', function ($req, $res) {
  * @route /:id
  */
 $router->put('/:id', function ($req, $res) {
-    return $res->send(["method" => "PUT"])->status(200);
+    return $res->send(["method" => "PUT"]);
 });
 
 /**
@@ -93,7 +124,7 @@ $router->put('/:id', function ($req, $res) {
  * @route /:id
  */
 $router->delete('/:id', function ($req, $res) {
-    return $res->send(["method" => "DELETE"])->status(200);
+    return $res->send(["method" => "DELETE"]);
 });
 
 // Start the router